fix circular export
