+++ package.json
@@ -33,7 +33,8 @@
         "build:test": "node scripts/test.mjs",
         "watch": "node ./scripts/watch.mjs",
         "watch:test": "npx nodemon",
-        "lint": "eslint && pnpm typecheck",
+        "eslint": "eslint --cache",
+        "lint": "pnpm eslint && pnpm typecheck",
         "package": "vsce package --no-dependencies",
         "vscode:prepublish": "pnpm build && pnpm lint",
         "publish": "vsce publish --no-dependencies",
